Address review: keep QuickJS asserts, fix the cache key, cover sys and fs
Four findings from the review, all verified before fixing rather than taken on trust.

The build type was RelWithDebInfo, which GCC expands to `-O2 -g -DNDEBUG`. That NDEBUG compiled out all 139 assert()s in quickjs.c, including `assert(p->ref_count > 0)` on the GC decref path — the cheapest detector for an over-released JSValue, and precisely the defect class this PR edits pr-automation.md to claim is now caught. Now Debug with the flags overridden to `-O1 -g`, which keeps traces readable and the suites under a second. Verified: quickjs.c.o references __assert_fail, and NDEBUG appears nowhere in build.ninja.

The CI cache keyed `_deps` on a constant. Trying to reproduce the predicted false green, LVGL did rebuild (804 files) — a fresh build.ninja with no .ninja_log treats every output as dirty, so the stale link does not currently occur. That is an accident of Ninja's behaviour rather than a guarantee, and the cache was also storing ~900 object files it rebuilds anyway. Now caches only lvgl-src and keys on hashFiles of both lv_conf.h files plus CMakeLists.txt, so it cannot outlive the configuration it was built for.

Two comments described code that does not exist. The host lv_conf.h listed LV_LOG_LEVEL as an override it never makes, while omitting LV_ASSERT_HANDLER, which is the one with a consequence: the board's `while(1);` would hang a CI job until the runner times out. test_reload's header attributed an EventBinding leak to the tracked counters, but those structs come from plain malloc (jsvm_core.cpp:197, :253) so only LSan sees them; the counters see the JSValues inside. Both rewritten to say what the checks actually do, including the case neither catches directly.

Also from the non-blocking notes:

- alloc_tracked offset the header by sizeof(size_t), handing back 8-mod-16 pointers — weaker alignment than either malloc or the IDF allocator promises, so a host pass could hide a target alignment fault. Padded to max_align_t, with the unchecked header addition now refused instead of wrapping. All three other header-offset sites were using sizeof(BlockHeader) and would have corrupted the heap.
- heap_caps_get_free_size and friends now return uint32_t, matching the ESP32 where size_t is 32-bit. This removes 7 -Wformat warnings that the printf attribute was raising against correct-on-target firmware; a clean build is now warning-free.
- Preferences::getString refused nothing and truncated silently. It now refuses, as the device does.
- bindings_sys.cpp and bindings_fs.cpp were compiled with no assertions over them. test_modules adds 22, covering the three host hooks (including NAN surfacing as null), backlight clamping, the pin surviving a stop/start pair, and the flash: prefix addressing a separate filesystem. This also gives the previously unused stub setters a caller.

69 assertions across four suites, 0 warnings, verified from a clean tree in an ubuntu:24.04 container. The trampoline mutation still produces heap-use-after-free in JS_CallInternal, and a doubled JS_FreeValue is caught too. No firmware source is modified.

+++ firmware/host-test/stubs/esp_heap_caps.h
@@ -39,9 +39,16 @@ void heap_caps_free(void *ptr);
 // Synthetic, and only plausible enough to keep the layer's diagnostics honest:
 // the free-size figures are a fixed notional pool minus what is outstanding.
 // Never assert on these in a test — assert on host_heap_live_bytes() instead.
-size_t heap_caps_get_free_size(uint32_t caps);
-size_t heap_caps_get_largest_free_block(uint32_t caps);
-size_t heap_caps_get_total_size(uint32_t caps);
+//
+// These return uint32_t rather than size_t deliberately. On the ESP32 size_t is
+// 32 bits, so the layer's `Serial.printf("%u bytes free", heap_caps_get_free_size(...))`
+// is correct there; declaring them size_t here (64-bit) would make the same
+// correct-on-target line produce a -Wformat warning on every host build, and the
+// only ways to silence that are editing firmware to suit the host or turning the
+// warning off. Matching the target's width is the honest fix.
+uint32_t heap_caps_get_free_size(uint32_t caps);
+uint32_t heap_caps_get_largest_free_block(uint32_t caps);
+uint32_t heap_caps_get_total_size(uint32_t caps);
 
 // QuickJS's usable_size hook must report 0 on the board (reporting real sizes
 // corrupts the heap under IDF poisoning; docs/engine-notes.md). The host copy of
